# [CHIA-1638] Pace block requests