--- a/chia/server/ws_connection.py
+++ b/chia/server/ws_connection.py
@@ -756,19 +756,9 @@ class WSChiaConnection:
         return capability in self.peer_capabilities
 
     def __hash__(self) -> int:
-        return hash(self.peer_node_id)
+        return hash(id(self))
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(object, WSChiaConnection):
             return False
-        return self.peer_node_id == other.peer_node_id
-
-    def __lt__(self, other: object) -> bool:
-        if not isinstance(object, WSChiaConnection):
-            return False
-        return self.peer_node_id < other.peer_node_id
-
-    def __gt__(self, other: object) -> bool:
-        if not isinstance(object, WSChiaConnection):
-            return False
-        return self.peer_node_id > other.peer_node_id
+        return id(self) == id(other)